Backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-dotenv
email-validator
langchain-google-genai
pdfplumber
python-docx
PyPDF2
pgvector
gunicorn
httpx
